build upload

+++ src/app/shared/navbar/navbar.component.ts
@@ -2,6 +2,10 @@ import { Component, OnInit, ElementRef } from '@angular/core';
 import { Location, LocationStrategy, PathLocationStrategy } from '@angular/common';
 import { Router } from '@angular/router';
 import { AuthService } from 'app/services/auth.service';
+import { AuthGuard } from '../auth.guard';
+import { ToastrService } from 'ngx-toastr';
+import { GetterSetterService } from '../../services/getter-setter.service';
+
 
 @Component({
     selector: 'app-navbar',
@@ -12,23 +16,30 @@ export class NavbarComponent implements OnInit {
     private toggleButton: any;
     private sidebarVisible: boolean;
     // tslint:disable-next-line:no-inferrable-types
-    signInButton: Boolean = false;
+    loggedIn: boolean;
+     userData: any;
 
     constructor(public location: Location,
         private element: ElementRef,
         private router: Router,
-        private authservice: AuthService) {
+        private authservice: AuthService,
+        public auth: AuthGuard,
+        private toastr: ToastrService,
+        private getSetService: GetterSetterService) {
         this.sidebarVisible = false;
     }
 
     ngOnInit() {
+        this.getSetService.getLoggedInStatus()
+        .subscribe((result) => {
+          this.loggedIn = result;
+        })
+      this.authservice.maintainStatus();
         const navbar: HTMLElement = this.element.nativeElement;
         this.toggleButton = navbar.getElementsByClassName('navbar-toggler')[0];
-        const userToken = localStorage.getItem('la_user_token');
-        if (userToken) {
-            this.signInButton = true;
-        }
+
     }
+
     sidebarOpen() {
         const toggleButton = this.toggleButton;
         const html = document.getElementsByTagName('html')[0];
@@ -77,7 +88,20 @@ export class NavbarComponent implements OnInit {
     }
 
     logOut() {
-        this.authservice.removeToken();
-        this.router.navigateByUrl('/login');
+        this.authservice.logout().subscribe(async res => {
+            if (res['status'] === true) {
+                this.getSetService.setLoggedInStatus(false);
+                this.authservice.removeToken();
+              this.toastr.success(res['message'], 'Spotlex');
+                  this.router.navigate(['/login']);
+                // this.route.navigate(['/order']);
+            } else if (res['status'] === false) {
+              this.toastr.error(res['message'], 'Spotlex');
+            } else {
+              this.toastr.error(res['message'], 'Spotlex');
+            }
+          }, (error) => {
+            this.toastr.error('error', error.error.message);
+          })
     }
 }
